Made length_check()
Moved test for the length of the path names to its own function instead of having in character_check(). It was only there to  limit the number of times it runs os.walk(), but now that needs to be repeated to do the renaming, I prefer it separate since it is a distinct test.