Update (still errors) unpack(), character_check()
Tested character_check() updates with full script and getting errors from the aip with spaces. 7z isn't unpacking it since it only sees -test, even once put it in single quotes. It did run with the full file path in double quotes separately on the terminal so try that next.