Do not delete original .tar.bz2 file
Do not delete the original AIPs in case the script has an error and needs to be run again. That way departments with large files don't have to make second copies, which can be slow.

Removed deletion from unpack() and added to tar() that the new AIPs are saved to a folder aptrust-aips within the AIPs directory to distinguish them from the original ARCHive AIPs.