Update paths for aptrust-aips folder throughout
Bags to be updated are in an aptrust-aips folder within the AIPs directory so they are not mixed with ARCHive tar and zip files. Added a variable aip_bag_path to include that folder and renamed the original variable from aip_bag to aip_bag_name to differentiate between the two better.