Change update_characters() to character_check()
Staff want to review names with impermissible characters instead of automatic renaming. All the names are saved to a log for review and the AIP(s) are moved to an error folder.

Now that no renaming is done, no longer need to account for a possibly renamed AIP with new_bag_path variable.