Move impermissible character log to error folder
Error folder isn't made until the first AIP bag is moved, so the log cannot be made directly in the error folder. Instead. it is made in the AIPs directory and then moved after the bag is moved. It is faster for staff review to have the log in the same place as the bag with the problems.